backend/pyproject.toml
[project]
name = "onramp-backend"
version = "0.1.0"
description = "Onramp document intake API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "pillow>=12.2.0",
    "pydantic-ai[anthropic]>=1.78.0",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "python-multipart>=0.0.24",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "fpdf2>=2.8.7",
    "httpx>=0.28.1",
    "mypy>=1.20.0",
    "pytest>=9.0.3",
    "ruff>=0.15.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
